Add Nokogiri XML parser
This one is similar to the LibXML parser, but uses the Nokogiri SAX parser under
the hood. This library is already used in a lot of web projects and is easier to
install in Windows than LibXML.
There is no direct dependency on the nokogiri gem, if the gem is not available
the code will only error in case it is used.