Add Nokogiri XML parser (#55)
This one is similar to the LibXML parser, but uses the Nokogiri SAX
parser under the hood. This library is already used in a lot of web
projects and is easier to install in Windows than LibXML.
There is no direct dependency on the nokogiri gem, if the gem is not
available the code will only error in case it is used.

Unfortunately, I could not find a way to enable the gem in the CI for
the Windows tasks. It passes on the regular Windows runners, but fails
on mingw/mswin, and I could not find a way to separate those two. If
someone knows how to fix that, or how to install/compile Nokogiri on the
latter two platforms, please go ahead and fix it.

This has been suggested/requested in
https://bugs.ruby-lang.org/issues/9379, all the way back in 2014.

---------

Co-authored-by: Herwin <herwinw@users.noreply.github.com>