Add XMLRPC::RackApplication (#56)
This is a simple Rack application that acts as an XMLRPC server. This
can be used in any Rack compatible webserver.

There are a few gems that provide a Rack-based XMLRPC server (rack-rpc,
xmlrpc-rack_server and threez-rack-rpc), but none of these have been
updated in the past 10 years.

Implementation has mostly been copied from ModRubyServer, tests have
been inspired by the WebrickServer tests

Co-authored-by: Herwin <herwinw@users.noreply.github.com>