Various layout cleanups in test_webrick_server
These remarks originate from the review of the tests of the CGI server,
but since most of that code was copied from this file, clean this up
here as well.
If we ever want to use shared logic for these files, it's preferable to
keep the differences to a minimum.