Add unit tests for XMLRPC::CGIServer (#60)
The implementation is mostly a copy-paste from the webrick server test,
including the code style.
Because this needs a CGI server, it still has a dependency on webrick to
run that server. This setup is a bit slow, on my system the overall
execution time of the tests jumps from 1 second to 7 seconds.

---------

Co-authored-by: Herwin <herwinw@users.noreply.github.com>
Co-authored-by: Sutou Kouhei <kou@cozmixng.org>